// src/app/shared/models/cart.ts
export interface Cart {
    tailorId: string;
    productId: string;
    name: string;
    image: string;
    fabric: string;
    category: string;
    price: number;
    quantity: number;
    total: number;
    delivery: number,
    duration: number;
    xsmallSize: number;
    smallSize: number;
    mediumSize: number;
    largeSize: number;
    xlargeSize: number;
    xxlargeSize: number;
}